Convert README to Markdown
Unlike other code that can be tested, before committed, I don't know of a way
to test this Markdown without pushing to github and seeing how github inteprets
it. So, here is another commit that may or may not be intereted in github the
way that I expect it to.