Wrap 'djurk' with basic Django project
When this repo was started, it was an app-level directory of the djurk app.  However, as it's possible to have an almost immediate demonstration for new usrs to download if we include a very basic shell of a django project, that's what we've done. We'll include instructions so it's easy to move the app from the Django project if the project isn't necessary.