Modify HIT to add generic foreign key
Djurk HITs are often associated with other Django Models. A generic
foreign key has been added to the HIT object to record that
relationship.

NOTE: The initial South migration has been updated. This would cause
production issues if this were live. However, as this is still alpha and
no one has yet put this in production, we updated the inital migration.