Remove use_sandbox parameter/build connection test
The use_sandbox parameter seemed like a good idea when the project was first started. However, allowing hits to be either from the sandbox or from the production system has caused more confusion that it has helped. When cached copies were saved, it was not obvious which environment that they came from. The sandbox can still be used. However, the sandbox host parameter is used explicity. This also simplifies the design and makes it more clean in several places.

Additionally, we had not test coverage for the settings parameters. Test settings have been added. Also, Deprecated ADMIN_MEDIA_PREFIX parameter was removed.